произведена дефрагментация
