+++ 12_2/12_2/12_2.vcxproj
@@ -150,10 +150,10 @@
     <Text Include="ReadMe.txt" />
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="stdafx.h" />
-    <ClInclude Include="targetver.h" />
+    <ClInclude Include="graphqueue.h" />
   </ItemGroup>
   <ItemGroup>
+    <ClCompile Include="graphqueue.cpp" />
     <ClCompile Include="main.cpp" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
